requirements.txt
django==5.0
djangorestframework==3.14.0
pre-commit==3.6.0
Pygments==2.17.2
